# Refactor/optimize validation