--- a/meta/Cargo.toml
+++ b/meta/Cargo.toml
@@ -16,7 +16,7 @@ include = ["Cargo.toml", "src/**/*", "src/grammar.rs", "_README.md", "LICENSE-*"
 [dependencies]
 maplit = "1.0"
 pest = { path = "../pest", version = "2.1.0" }
-lazy_static = "1.4.0"
+once_cell = "1.8.0"
 
 [badges]
 codecov = { repository = "pest-parser/pest" }