--- a/meta/src/lib.rs
+++ b/meta/src/lib.rs
@@ -9,6 +9,9 @@
 
 #![allow(clippy::range_plus_one)]
 
+#[macro_use]
+extern crate lazy_static;
+
 extern crate maplit;
 #[cfg(test)]
 #[macro_use]
@@ -41,7 +44,8 @@ where
 
 #[doc(hidden)]
 pub static UNICODE_PROPERTY_NAMES: &[&str] = &[
-    /* BINARY */ "ALPHABETIC",
+    /* BINARY */
+    "ALPHABETIC",
     "BIDI_CONTROL",
     "CASE_IGNORABLE",
     "CASED",
@@ -93,7 +97,8 @@ pub static UNICODE_PROPERTY_NAMES: &[&str] = &[
     "WHITE_SPACE",
     "XID_CONTINUE",
     "XID_START",
-    /* CATEGORY */ "CASED_LETTER",
+    /* CATEGORY */
+    "CASED_LETTER",
     "CLOSE_PUNCTUATION",
     "CONNECTOR_PUNCTUATION",
     "CONTROL",