@tomtau 2022-07-20
At least those early benchmark results looked interesting. The current main bottlenecks are elsewhere though: e.g. https://github.com/pest-parser/pest/issues/560