@NoahTheDuke 2022-07-20
Yeah this was a fun romp and I think works be beneficial overall, but I want convinced that it was worth the trade off to readability. I stopped using Pest regularly shortly after this which is why I closed it but I'd be willing to dust it off and see if it still provides "dramatic" improvements.